backend/requirements.txt
alembic==1.16.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
distro==1.9.0
fastapi==0.128.0
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.16.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.1.1
sgmllib3k==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.50.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
